library 1.0

+++ htab_hash_function.c
@@ -1,7 +1,7 @@
 #include "htab.h"
 #include "stdint.h"
 
-unsigned int htab_hash_fun(const char *str) {
+size_t htab_hash_fun(const char *str) {
     uint32_t h = 0;     // musí mít 32 bitů
     const unsigned char *p;
     for(p = (const unsigned char*) str; *p != '\0'; p++) {
